create Cardlist and Styling
