+++ components/CardList/index.js
@@ -1,44 +1,32 @@
+import useSWR from "swr";
+import ListCard from "../ListCard";
 import styled from "styled-components";
-import Image from "next/image";
 
 export default function CardList() {
+  const { data, error } = useSWR("/api/db");
+  console.log(data);
+
   return (
-    <>
-      <StyledCard>
-        <StyledCardHeader>Card 1</StyledCardHeader>
-        <StyledCardImage
-          src="/pmlogo.png"
-          alt="PawfectMatch"
-          width={200}
-          height={50}
-        />
-      </StyledCard>
-    </>
+    <StyledCardList>
+      {data ? (
+        data.map((dog) => (
+          <div key={dog.id}>
+            <ListCard name={dog.name} img={dog.image_link} />
+          </div>
+        ))
+      ) : (
+        <div>Loading...</div>
+      )}
+    </StyledCardList>
   );
 }
 
-const StyledCard = styled.article`
+const StyledCardList = styled.section`
   display: flex;
   flex-direction: column;
   align-items: center;
-  width: 90vw;
-  height: 50vh;
-  background-color: #fff;
-  border-radius: 30px;
-  border: 1px solid #000;
-`;
-
-const StyledCardHeader = styled.header`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 50px;
-  background-color: #fff;
-  border-radius: 30px;
-`;
-
-const StyledCardImage = styled(Image)`
   width: 100%;
   height: 100%;
+  gap: 20px;
+  margin-top: 20px;
 `;
